pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Andrey <lutecequantum@gmail.com>"]
packages = [
	{ include = 'page_analyzer' },
]

[tool.poetry.dependencies]
python = "^3.8.1"
Flask = "^2.3.2"
flake8 = "^6.0.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
validators = "^0.20.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
